# gpt/openapi.yaml
openapi: 3.1.0
info:
  title: Resume Builder API
  description: Backend API for the Resume Builder Custom GPT. Handles module storage, JD matching, and LaTeX compilation.
  version: 2.0.0
servers:
  - url: https://resume-builder-production-229e.up.railway.app
    description: Production server (Railway)
paths:
...
  /api/modules:
    post:
      operationId: storeModules
      summary: Store parsed resume modules and bullets
      description: Persist structured resume modules and their bullet-level evidence units into the database.
      x-openai-isConsequential: false
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                modules:
                  type: array
                  description: Experience-level modules
                  items:
                    type: object
                    required: [module_id, type, section, organization, title, date_range, context_tags, base_priority]
                    properties:
                      module_id:
                        type: string
                      type:
                        type: string
                        enum: [experience, project, education, leadership, award, certification]
                      section:
                        type: string
                        enum: [experience, education, projects, leadership, skills, awards]
                      organization:
                        type: string
                      title:
                        type: string
                      date_range:
                        type: string
                      location:
                        type: string
                        description: "City and country of the experience (e.g. 'Beijing, China' or '北京，中国'). Optional."
                      context_tags:
                        type: array
                        items:
                          type: string
                      base_priority:
                        type: number
                      source_type:
                        type: string
                        default: parsed
                bullets:
                  type: array
                  description: Bullet-level evidence units
                  items:
                    type: object
                    required: [bullet_id, parent_module_id, raw_fact, normalized_fact, evidence_tags, skill_tags, role_fit_tags, strength_score]
                    properties:
                      bullet_id:
                        type: string
                      parent_module_id:
                        type: string
                      raw_fact:
                        type: string
                      normalized_fact:
                        type: object
                        properties:
                          action:
                            type: string
                          object:
                            type: string
                          method:
                            type: array
                            items:
                              type: string
                          output:
                            type: string
                          metric:
                            type: string
                            nullable: true
                      evidence_tags:
                        type: array
                        items:
                          type: string
                      skill_tags:
                        type: array
                        items:
                          type: string
                      role_fit_tags:
                        type: array
                        items:
                          type: string
                      strength_score:
                        type: object
                        properties:
                          clarity:
                            type: number
                          quantification:
                            type: number
                          brand_signal:
                            type: number
                          transferability:
                            type: number
                      rewrite_candidates:
                        type: array
                        items:
                          type: string
                        default: []
      responses:
        "200":
          description: Modules stored successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  modules_stored:
                    type: integer
                  bullets_stored:
                    type: integer
    get:
      operationId: listModules
      summary: List all stored resume modules with bullets
      description: Retrieve all resume modules and their associated bullets from the database.
      responses:
        "200":
          description: List of modules
          content:
            application/json:
              schema:
                type: object
                properties:
                  modules:
                    type: array
                    items:
                      type: object
                  count:
                    type: integer
  /api/modules/{id}:
    get:
      operationId: getModule
      summary: Get a single module with its bullets
      description: Retrieve one resume module by ID, including all associated bullets.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Module with bullets
          content:
            application/json:
              schema:
                type: object
                properties:
                  module_id:
                    type: string
                  organization:
                    type: string
                  title:
                    type: string
                  section:
                    type: string
                  date_range:
                    type: string
                  location:
                    type: string
                  context_tags:
                    type: array
                    items:
                      type: string
                  base_priority:
                    type: number
                  bullets:
                    type: array
                    items:
                      type: object
                      properties:
                        bullet_id:
                          type: string
                        raw_fact:
                          type: string
                        evidence_tags:
                          type: array
                          items:
                            type: string
        "404":
          description: Module not found
    delete:
      operationId: deleteModule
      summary: Delete a stored resume module
      description: Permanently delete a module and all its bullets. Use when the user explicitly asks to remove a module from the library.
      x-openai-isConsequential: true
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Module deleted
        "404":
          description: Module not found
  /api/jd:
    post:
      operationId: storeJd
      summary: Store a structured job description
      description: Persist a parsed JD with requirements, signals, and style constraints.
      x-openai-isConsequential: false
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [job_id, raw_text, meta, hard_requirements, soft_requirements, preferred_signals, domain_tags, evidence_targets, style_constraints]
              properties:
                job_id:
                  type: string
                raw_text:
                  type: string
                meta:
                  type: object
                  properties:
                    company:
                      type: string
                    team:
                      type: string
                    location:
                      type: string
                    role_title:
                      type: string
                    language:
                      type: string
                      default: en
                    seniority:
                      type: string
                      default: student_intern
                hard_requirements:
                  type: array
                  items:
                    type: string
                soft_requirements:
                  type: array
                  items:
                    type: string
                preferred_signals:
                  type: array
                  items:
                    type: string
                domain_tags:
                  type: array
                  items:
                    type: string
                evidence_targets:
                  type: array
                  items:
                    type: object
                    properties:
                      signal:
                        type: string
                      examples:
                        type: array
                        items:
                          type: string
                      priority:
                        type: number
                style_constraints:
                  type: object
                  properties:
                    resume_language:
                      type: string
                      default: en
                    bullet_style:
                      type: string
                      default: action_first
                    quant_preference:
                      type: string
                      default: high
                    tone:
                      type: string
                      default: professional_student
      responses:
        "200":
          description: JD stored successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  job_id:
                    type: string
                  role:
                    type: string
                  company:
                    type: string
    get:
      operationId: listJds
      summary: List all stored job descriptions
      description: Retrieve all stored JD records.
      responses:
        "200":
          description: List of JDs
          content:
            application/json:
              schema:
                type: object
                properties:
                  jds:
                    type: array
                    items:
                      type: object
                  count:
                    type: integer
...
  /api/children/{cid}/link-jd:
    post:
      operationId: linkChildJd
      summary: Link an existing child module to an additional JD
      description: Associate a child module with another JD for reuse across similar positions.
      x-openai-isConsequential: false
      parameters:
        - name: cid
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [job_id]
              properties:
                job_id:
                  type: string
      responses:
        "200":
          description: JD linked
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
...